requirements.txt

discord.py
aiohttp
flask
requests
psutil
python-dotenv
aiohttp
discord.py
flask
psutil
python-dotenv
requests
